- вся бизнес - логика из контроллеров перенесена в сервисы. Сервисы внедрены в контроллеры при помощи зависимостей;
