+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -2,35 +2,34 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.ValidateException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.FilmService;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.HashMap;
+
+import java.util.List;
 
 
 @RestController
 @RequestMapping("/films")
 @Slf4j
 public class FilmController {
-    protected HashMap<Long, Film> filmStorage = new HashMap<>();
+    private final FilmService filmService;
 
 
-    public FilmController() {
+    @Autowired
+    public FilmController(FilmService filmService) {
+        this.filmService = filmService;
     }
 
     @PostMapping
     public @ResponseBody Film create(@RequestBody Film film) throws ValidateException {
 
         validateNewFilm(film);
-        film.setId(generateNewId());
-        if (film.getName() == null) {
-            log.error("Название фильма не может быть пустым");
-            throw new ValidateException("Название фильма не может быть пустым");
-        }
-        filmStorage.put(film.getId(), film);
+        filmService.createFilm(film, this);
         return film;
     }
 
@@ -39,13 +38,7 @@ public class FilmController {
 
         validateNewFilm(film);
 
-        if (film.getId() == null || !filmStorage.containsKey(film.getId())) {
-            log.error("id  не может быть null");
-            throw new ValidateException("id  не может быть null");
-        } else {
-            filmStorage.put(film.getId(), film);
-        }
-        return film;
+        return filmService.updateFilm(film);
 
     }
 
@@ -72,23 +65,9 @@ public class FilmController {
     }
 
     @GetMapping
-    public @ResponseBody ArrayList<Film> getAllFilm() {
-        return new ArrayList<>(filmStorage.values());
+    public @ResponseBody List<Film> getAllFilm() {
+        return filmService.getAllFilms();
     }
 
-    public long getMaxId() {
-        long maxId = 0;
-        for (long id : filmStorage.keySet()) {
-            if (id > maxId) {
-                maxId = id;
-            }
-        }
-        return maxId;
-    }
 
-    public long generateNewId() {
-        long maxId = getMaxId();
-        long newId = maxId + 1;
-        return newId;
-    }
 }
