+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,45 +1,40 @@
 package ru.yandex.practicum.filmorate.controller;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import ru.yandex.practicum.filmorate.exception.ValidateException;
 import lombok.extern.slf4j.Slf4j;
 import ru.yandex.practicum.filmorate.model.User;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.storage.UserService;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.List;
 
 @RestController
 @RequestMapping("/users")
 @Slf4j
 public class UserController {
-    protected HashMap<Long, User> userStorage = new HashMap<>();
 
-    public UserController() {
+    private final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
     }
 
 
     @PostMapping
     public @ResponseBody User create(@RequestBody User user) throws ValidateException {
         validateNewUser(user);
-        user.setId(generateNewId());
-        if (user.getName() == null) {
-            user.setName(user.getLogin());
-        }
-        userStorage.put(user.getId(), user);
+        userService.createUser(user);
         return user;
     }
+
 
     @PutMapping
     public @ResponseBody User update(@RequestBody User user) throws ValidateException {
         validateNewUser(user);
-        if (user.getId() == null || !userStorage.containsKey(user.getId())) {
-            log.error("id  не может быть null");
-            throw new ValidateException("id  не может быть null");
-        } else {
-            userStorage.put(user.getId(), user);
-        }
-        return user;
+        return userService.updateUser(user);
     }
 
 
@@ -67,23 +62,9 @@ public class UserController {
     }
 
     @GetMapping
-    public @ResponseBody ArrayList<User> getAllUsers() {
-        return new ArrayList<>(userStorage.values());
+    public @ResponseBody List<User> getAllUsers() {
+        return userService.getAllUsers();
     }
 
-    private long getMaxId() {
-        long maxId = 0;
-        for (long id : userStorage.keySet()) {
-            if (id > maxId) {
-                maxId = id;
-            }
-        }
-        return maxId;
-    }
 
-    public long generateNewId() {
-        long maxId = getMaxId();
-        long newId = maxId + 1;
-        return newId;
-    }
 }
