+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,0 +1,4 @@
+package ru.yandex.practicum.filmorate.storage;
+
+public interface FilmStorage {
+}
